+++ CHANGES.md
@@ -9,6 +9,40 @@
 
 <!-- towncrier release notes start -->
 
+## 1.0.0a3 (2023-11-24)
+
+
+### New features:
+
+- Control creation of "Authors" folder via registry configuration [@ericof] #34
+- Add control panel to configure Blog Settings [@ericof] #36
+- Additional information returned on /@authors endpoint [@ericof] #40
+
+
+### Internal:
+
+- Increase test coverage of the package [@ericof] #38
+
+
+## 1.0.0a2 (2023-11-23)
+
+
+### New features:
+
+- Initial Package Structure for Plone 6 [@jonaspiterek] #11
+- Blog and Author should be listed on Plone Navigation [@jonaspiterek] #13
+- Implement i18n support [@ericof] #17
+- Implement `collective.blog.blog_info` behavior [@ericof] #20
+- Disable allowing to add Author globally. Fix GS titles [@jonaspiterek] #21
+- Add index and metadata for blog post and blog author, Add test for blog author [@jonaspiterek] #28
+- Implement collective.blog: Add Author, collective.blog: Add Blog, collective.blog: Add Post permissions [@ericof] #30
+- Create folder authors inside a blog [@jonaspiterek] #32
+
+
+### Documentation:
+
+- Improve README.md [@ericof] #15
+
 
 ## 1.0a1 (2016-02-14)